pheromone testing

+++ AntSim/src/at/antSim/objectsKI/Queen.java
@@ -15,7 +15,7 @@ import at.antSim.objectsPhysic.basics.PhysicsObject;
  *
  */
 public class Queen  implements Feedable {
-	private int hp;
+	private float hp;
 	private int hunger;
 	private int eggLayingCost = 1;
 	private int eggLayingSpeed = 1;
@@ -38,11 +38,11 @@ public class Queen  implements Feedable {
 		}
 	}
 
-	public int getHp() {
+	public float getHp() {
 		return hp;
 	}
 
-	public void setHp(int hp) {
+	public void setHp(float hp) {
 		this.hp = hp;
 	}
